# [libc++] Optimize ranges::copy_backward for vector<bool>::iterator