--- a/libcxx/test/std/algorithms/alg.modifying.operations/alg.copy/ranges.copy_backward.pass.cpp
+++ b/libcxx/test/std/algorithms/alg.modifying.operations/alg.copy/ranges.copy_backward.pass.cpp
@@ -348,9 +348,11 @@ constexpr bool test() {
 #if TEST_STD_VER >= 23
   { // Test vector<bool>::iterator optimization
     assert(test_vector_bool(8));
-    assert(test_vector_bool(16));
+    assert(test_vector_bool(19));
     assert(test_vector_bool(32));
+    assert(test_vector_bool(49));
     assert(test_vector_bool(64));
+    assert(test_vector_bool(199));
     assert(test_vector_bool(256));
   }
 #endif