As a follow-up to #121013 (which focused on `std::ranges::copy`), this PR optimizes the performance of `std::ranges::copy_backward` for `vector<bool>::iterator`, addressing a subtask outlined in issue #64038. The optimizations yield performance improvements of up to **2000x** for aligned copies and **60x** for unaligned copies. 

- Aligned copy
```
------------------------------------------------------------------------------------
Benchmark                                        Before        After     Improvement
------------------------------------------------------------------------------------
bm_ranges_copy_backward_vb_aligned/8            10.7 ns      1.74 ns            6x
bm_ranges_copy_backward_vb_aligned/64           78.7 ns      3.07 ns           26x
bm_ranges_copy_backward_vb_aligned/512           617 ns      2.77 ns          223x
bm_ranges_copy_backward_vb_aligned/4096         4902 ns      5.91 ns          829x
bm_ranges_copy_backward_vb_aligned/32768       43111 ns      40.1 ns         1075x
bm_ranges_copy_backward_vb_aligned/65536       80592 ns      71.9 ns         1121x
bm_ranges_copy_backward_vb_aligned/102400     129705 ns       121 ns         1072x
bm_ranges_copy_backward_vb_aligned/106496     134432 ns      85.3 ns         1576x
bm_ranges_copy_backward_vb_aligned/110592     140523 ns      68.5 ns         2051x
bm_ranges_copy_backward_vb_aligned/114688     139851 ns      70.8 ns         1975x
bm_ranges_copy_backward_vb_aligned/118784     146393 ns      73.3 ns         1997x
bm_ranges_copy_backward_vb_aligned/122880     152805 ns      73.5 ns         2079x
bm_ranges_copy_backward_vb_aligned/126976     154205 ns      76.8 ns         2008x
bm_ranges_copy_backward_vb_aligned/131072     160370 ns       136 ns         1179x
bm_ranges_copy_backward_vb_aligned/135168     165428 ns       161 ns         1028x
bm_ranges_copy_backward_vb_aligned/139264     169763 ns       109 ns         1557x
bm_ranges_copy_backward_vb_aligned/143360     175159 ns      86.8 ns         2018x
bm_ranges_copy_backward_vb_aligned/147456     178807 ns      88.1 ns         2030x
bm_ranges_copy_backward_vb_aligned/151552     184829 ns      92.1 ns         2007x
bm_ranges_copy_backward_vb_aligned/155648     193208 ns      90.8 ns         2128x
bm_ranges_copy_backward_vb_aligned/159744     198470 ns      93.4 ns         2125x
bm_ranges_copy_backward_vb_aligned/163840     205261 ns       170 ns         1207x
bm_ranges_copy_backward_vb_aligned/167936     210831 ns       198 ns         1064x
bm_ranges_copy_backward_vb_aligned/172032     215258 ns       132 ns         1630x
bm_ranges_copy_backward_vb_aligned/176128     219453 ns       111 ns         1977x
bm_ranges_copy_backward_vb_aligned/180224     227048 ns       110 ns         2064x
bm_ranges_copy_backward_vb_aligned/184320     230324 ns       115 ns         2002x
bm_ranges_copy_backward_vb_aligned/188416     234305 ns       121 ns         1936x
bm_ranges_copy_backward_vb_aligned/192512     239388 ns       127 ns         1884x
bm_ranges_copy_backward_vb_aligned/196608     246586 ns       214 ns         1152x
bm_ranges_copy_backward_vb_aligned/200704     249973 ns       264 ns          946x
bm_ranges_copy_backward_vb_aligned/204800     259443 ns       221 ns         1173x
```

- Unaligned copy

```
---------------------------------------------------------------------------------------
Benchmark                                           Before           After  Improvement
---------------------------------------------------------------------------------------
bm_ranges_copy_backward_vb_unaligned/8             11.0 ns         9.49 ns         1.2x
bm_ranges_copy_backward_vb_unaligned/64            95.9 ns         8.18 ns          12x
bm_ranges_copy_backward_vb_unaligned/512            720 ns         18.8 ns          38x
bm_ranges_copy_backward_vb_unaligned/4096          5871 ns          103 ns          57x
bm_ranges_copy_backward_vb_unaligned/32768        45974 ns          741 ns          62x
bm_ranges_copy_backward_vb_unaligned/262144      374481 ns         6100 ns          61x
bm_ranges_copy_backward_vb_unaligned/1048576    1531217 ns        25406 ns          60x
```